+++ strategies/vim_session.sh
@@ -4,7 +4,7 @@
 #
 # Restores a vim session from 'Session.vim' file, if it exists.
 # If 'Session.vim' does not exist, it falls back to invoking the original
-# command.
+# command (withouth the `-S` flag).
 
 ORIGINAL_COMMAND="$1"
 DIRECTORY="$2"
@@ -13,9 +13,18 @@ vim_session_file_exists() {
 	[ -e "${DIRECTORY}/Session.vim" ]
 }
 
+original_command_contains_session_flag() {
+	[[ "$ORIGINAL_COMMAND" =~ "-S" ]]
+}
+
 main() {
 	if vim_session_file_exists; then
 		echo "vim -S"
+	elif original_command_contains_session_flag; then
+		# Session file does not exist, yet the orignal vim command contains
+		# session flag `-S`. This will cause an error, so we're falling back to
+		# starting plain vim.
+		echo "vim"
 	else
 		echo "$ORIGINAL_COMMAND"
 	fi